[Feat] 병합 오류 수정 및 변수명 변경
Player_ID -> Player_Name
Party_ID -> Party_Name

+++ Assets/02. Scripts/UI/UIStatStore.cs
@@ -175,7 +175,7 @@ public class UIStatStore : UIPopup
                 AppliedHpLevel,
                 hpIncrease,
                 AppliedDamageLevel,
-                damageIncrease,
+                (int)damageIncrease,
                 RemainingPoint
             );
         }
